+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement daily monitoring email for failed tasks / collectors
 - Implement nist_cvss_validation in Flaw API (OSIDB-1006)
 - Implement additional tracker validations (OSIDB-787)
+- Validate NIST RH CVSS feedback loop (OSIDB-334)
 
 ### Changed
 - Change article link validation to be blocking (OSIDB-1060)